+++ test/services/dumper/dumper-sequelize.js
@@ -4,6 +4,7 @@ const fs = require('fs');
 
 const simpleModel = require('../../expected/sequelize/db-analysis-output/customers.json');
 const belongsToModel = require('../../expected/sequelize/db-analysis-output/addresses.json');
+const otherAssociationsModel = require('../../expected/sequelize/db-analysis-output/users.json');
 
 const Dumper = require('../../../services/dumper');
 
@@ -39,6 +40,14 @@ describe('Dumper > sequelize', () => {
     await dumper.dump(belongsToModel);
     const generatedFile = fs.readFileSync('./test/output/sequelize/models/addresses.js', 'utf8');
     const expectedFile = fs.readFileSync('./test/expected/sequelize/dumper-output/addresses.js.expected', 'utf-8');
+
+    expect(generatedFile).to.equals(expectedFile);
+  });
+
+  it('generate a model file with hasMany, hasOne and belongsToMany', async () => {
+    await dumper.dump(otherAssociationsModel);
+    const generatedFile = fs.readFileSync('./test/output/sequelize/models/users.js', 'utf8');
+    const expectedFile = fs.readFileSync('./test/expected/sequelize/dumper-output/users.js.expected', 'utf-8');
 
     expect(generatedFile).to.equals(expectedFile);
   });